requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cryptography==2.3
Django==2.0.7
django-webpush==0.2.5
http-ece==1.0.5
idna==2.7
py-vapid==1.4.0
pycparser==2.18
pytz==2018.5
pywebpush==1.6.0
requests==2.19.1
six==1.11.0
urllib3==1.23
